Fix printf format string warning
There is currently an assumption in the Cyclone serializers that an enum maps to
an (unsigned) int and than an int is 32-bits. Because of that, printing an enum with a
format string of PRIu32 will work fine, but there should be a cast to uint32_t
nonetheless.

+++ src/core/ddsi/src/ddsi_plist.c
@@ -1778,7 +1778,7 @@ static bool print_generic1 (char * restrict *buf, size_t * restrict bufsize, con
       }
       case XE1: case XE2: case XE3: { /* enum */
         enum xe3_prototype const * const x = deser_generic_src (src, &srcoff, plist_alignof (enum xe3_prototype));
-        if (!prtf (buf, bufsize, "%s%"PRIu32, sep, *x))
+        if (!prtf (buf, bufsize, "%s%"PRIu32, sep, (uint32_t) *x))
           return false;
         srcoff += sizeof (*x);
         break;
